requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==24.3
django-storages==1.14.4
boto3==1.35.76
Pillow==10.4.0
minio==7.2.7
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
xhtml2pdf==0.2.16
django-cors-headers==4.4.0
